Refactored package structure and optimised imports
