+++ src/main/java/dev/tycho/stonks/database/AccountLinkDaoImpl.java
@@ -3,13 +3,10 @@ package dev.tycho.stonks.database;
 import com.j256.ormlite.dao.BaseDaoImpl;
 import com.j256.ormlite.stmt.QueryBuilder;
 import com.j256.ormlite.support.ConnectionSource;
-import dev.tycho.stonks.model.*;
-import dev.tycho.stonks.model.accountvisitors.IAccountVisitor;
 import dev.tycho.stonks.model.accountvisitors.ReturningAccountVisitor;
+import dev.tycho.stonks.model.core.*;
 
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
 
 public class AccountLinkDaoImpl extends BaseDaoImpl<AccountLink, Integer> {
     public AccountLinkDaoImpl(ConnectionSource connectionSource) throws SQLException {
