+++ src/main/java/dev/tycho/stonks/database/CompanyAccountDao.java
@@ -1,7 +1,7 @@
 package dev.tycho.stonks.database;
 
 import com.j256.ormlite.dao.Dao;
-import dev.tycho.stonks.model.CompanyAccount;
+import dev.tycho.stonks.model.core.CompanyAccount;
 
 public interface CompanyAccountDao extends Dao<CompanyAccount, Integer> {
 }
