+++ src/main/java/dev/tycho/stonks/database/CompanyAccountDaoImpl.java
@@ -2,7 +2,7 @@ package dev.tycho.stonks.database;
 
 import com.j256.ormlite.dao.BaseDaoImpl;
 import com.j256.ormlite.support.ConnectionSource;
-import dev.tycho.stonks.model.CompanyAccount;
+import dev.tycho.stonks.model.core.CompanyAccount;
 
 import java.sql.SQLException;
