+++ src/main/java/dev/tycho/stonks/database/HoldingDao.java
@@ -1,8 +1,8 @@
 package dev.tycho.stonks.database;
 
 import com.j256.ormlite.dao.Dao;
-import dev.tycho.stonks.model.Holding;
-import dev.tycho.stonks.model.Member;
+import dev.tycho.stonks.model.core.Holding;
+import dev.tycho.stonks.model.core.Member;
 
 public interface HoldingDao extends Dao<Holding, Integer> {
     boolean memberHasHoldings(Member member);
