+++ src/main/java/dev/tycho/stonks/database/HoldingDaoImpl.java
@@ -2,8 +2,8 @@ package dev.tycho.stonks.database;
 
 import com.j256.ormlite.dao.BaseDaoImpl;
 import com.j256.ormlite.support.ConnectionSource;
-import dev.tycho.stonks.model.Holding;
-import dev.tycho.stonks.model.Member;
+import dev.tycho.stonks.model.core.Holding;
+import dev.tycho.stonks.model.core.Member;
 
 import java.sql.SQLException;
