+++ src/main/java/dev/tycho/stonks/database/MemberDao.java
@@ -1,9 +1,9 @@
 package dev.tycho.stonks.database;
 
 import com.j256.ormlite.dao.Dao;
-import dev.tycho.stonks.model.Company;
-import dev.tycho.stonks.model.Member;
-import dev.tycho.stonks.model.Role;
+import dev.tycho.stonks.model.core.Company;
+import dev.tycho.stonks.model.core.Member;
+import dev.tycho.stonks.model.core.Role;
 import org.bukkit.entity.Player;
 
 import java.sql.SQLException;
