+++ src/main/java/dev/tycho/stonks/database/TransactionDaoImpl.java
@@ -4,8 +4,7 @@ import com.j256.ormlite.dao.BaseDaoImpl;
 import com.j256.ormlite.stmt.QueryBuilder;
 import com.j256.ormlite.support.ConnectionSource;
 import dev.tycho.stonks.logging.Transaction;
-import dev.tycho.stonks.model.AccountLink;
-import org.bukkit.entity.Player;
+import dev.tycho.stonks.model.core.AccountLink;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
