+++ src/main/java/dev/tycho/stonks/gui/AccountSelectorGui.java
@@ -1,9 +1,9 @@
 package dev.tycho.stonks.gui;
 
-import com.j256.ormlite.dao.ForeignCollection;
 import dev.tycho.stonks.managers.DatabaseManager;
-import dev.tycho.stonks.model.*;
-import dev.tycho.stonks.model.accountvisitors.ReturningAccountVisitor;
+import dev.tycho.stonks.model.core.Account;
+import dev.tycho.stonks.model.core.AccountLink;
+import dev.tycho.stonks.model.core.Company;
 import dev.tycho.stonks.util.Util;
 import fr.minuskube.inv.ClickableItem;
 import fr.minuskube.inv.InventoryManager;
@@ -12,13 +12,9 @@ import fr.minuskube.inv.content.InventoryContents;
 import fr.minuskube.inv.content.InventoryProvider;
 import fr.minuskube.inv.content.Pagination;
 import fr.minuskube.inv.content.SlotIterator;
-import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
 import java.util.function.Consumer;
 
 public class AccountSelectorGui implements InventoryProvider {
