+++ src/main/java/dev/tycho/stonks/gui/CompanyInfoGui.java
@@ -1,13 +1,13 @@
 package dev.tycho.stonks.gui;
 
+import dev.tycho.stonks.managers.DatabaseManager;
+import dev.tycho.stonks.model.core.Company;
+import dev.tycho.stonks.util.Util;
 import fr.minuskube.inv.ClickableItem;
 import fr.minuskube.inv.InventoryManager;
 import fr.minuskube.inv.SmartInventory;
 import fr.minuskube.inv.content.InventoryContents;
 import fr.minuskube.inv.content.InventoryProvider;
-import dev.tycho.stonks.model.Company;
-import dev.tycho.stonks.managers.DatabaseManager;
-import dev.tycho.stonks.util.Util;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
