+++ src/main/java/dev/tycho/stonks/gui/CompanyListGui.java
@@ -1,5 +1,8 @@
 package dev.tycho.stonks.gui;
 
+import dev.tycho.stonks.managers.DatabaseManager;
+import dev.tycho.stonks.model.core.Company;
+import dev.tycho.stonks.util.Util;
 import fr.minuskube.inv.ClickableItem;
 import fr.minuskube.inv.InventoryManager;
 import fr.minuskube.inv.SmartInventory;
@@ -7,9 +10,6 @@ import fr.minuskube.inv.content.InventoryContents;
 import fr.minuskube.inv.content.InventoryProvider;
 import fr.minuskube.inv.content.Pagination;
 import fr.minuskube.inv.content.SlotIterator;
-import dev.tycho.stonks.model.Company;
-import dev.tycho.stonks.managers.DatabaseManager;
-import dev.tycho.stonks.util.Util;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
