+++ src/main/java/dev/tycho/stonks/gui/ConfirmationGui.java
@@ -1,8 +1,6 @@
 package dev.tycho.stonks.gui;
 
 import dev.tycho.stonks.managers.DatabaseManager;
-import dev.tycho.stonks.model.AccountType;
-import dev.tycho.stonks.model.Company;
 import dev.tycho.stonks.util.Util;
 import fr.minuskube.inv.ClickableItem;
 import fr.minuskube.inv.InventoryManager;
