+++ src/main/java/dev/tycho/stonks/gui/InviteListGui.java
@@ -1,5 +1,9 @@
 package dev.tycho.stonks.gui;
 
+import dev.tycho.stonks.managers.DatabaseManager;
+import dev.tycho.stonks.model.core.Company;
+import dev.tycho.stonks.model.core.Member;
+import dev.tycho.stonks.util.Util;
 import fr.minuskube.inv.ClickableItem;
 import fr.minuskube.inv.InventoryManager;
 import fr.minuskube.inv.SmartInventory;
@@ -7,10 +11,6 @@ import fr.minuskube.inv.content.InventoryContents;
 import fr.minuskube.inv.content.InventoryProvider;
 import fr.minuskube.inv.content.Pagination;
 import fr.minuskube.inv.content.SlotIterator;
-import dev.tycho.stonks.model.Company;
-import dev.tycho.stonks.model.Member;
-import dev.tycho.stonks.managers.DatabaseManager;
-import dev.tycho.stonks.util.Util;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
