+++ src/main/java/dev/tycho/stonks/gui/TransactionHistoryGui.java
@@ -2,10 +2,10 @@ package dev.tycho.stonks.gui;
 
 import dev.tycho.stonks.logging.Transaction;
 import dev.tycho.stonks.managers.DatabaseManager;
-import dev.tycho.stonks.model.AccountLink;
-import dev.tycho.stonks.model.CompanyAccount;
-import dev.tycho.stonks.model.HoldingsAccount;
 import dev.tycho.stonks.model.accountvisitors.IAccountVisitor;
+import dev.tycho.stonks.model.core.AccountLink;
+import dev.tycho.stonks.model.core.CompanyAccount;
+import dev.tycho.stonks.model.core.HoldingsAccount;
 import dev.tycho.stonks.util.Util;
 import fr.minuskube.inv.ClickableItem;
 import fr.minuskube.inv.InventoryManager;
