+++ src/main/java/dev/tycho/stonks/logging/Transaction.java
@@ -2,8 +2,7 @@ package dev.tycho.stonks.logging;
 
 import com.Acrobot.ChestShop.ORMlite.table.DatabaseTable;
 import com.j256.ormlite.field.DatabaseField;
-import dev.tycho.stonks.model.AccountLink;
-import org.bukkit.Material;
+import dev.tycho.stonks.model.core.AccountLink;
 
 import java.sql.Timestamp;
 import java.util.Calendar;
