+++ src/main/java/dev/tycho/stonks/managers/MessageManager.java
@@ -1,7 +1,7 @@
 package dev.tycho.stonks.managers;
 
 import dev.tycho.stonks.Stonks;
-import dev.tycho.stonks.model.Member;
+import dev.tycho.stonks.model.core.Member;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
