+++ src/main/java/dev/tycho/stonks/managers/ShopManager.java
@@ -7,8 +7,8 @@ import com.Acrobot.ChestShop.UUIDs.NameManager;
 import com.j256.ormlite.stmt.QueryBuilder;
 import dev.tycho.stonks.Stonks;
 import dev.tycho.stonks.logging.Transaction;
-import dev.tycho.stonks.model.*;
 import dev.tycho.stonks.model.accountvisitors.IAccountVisitor;
+import dev.tycho.stonks.model.core.*;
 import org.bukkit.ChatColor;
 import org.bukkit.event.EventHandler;
 
@@ -149,7 +149,7 @@ public class ShopManager extends SpigotModule {
     @EventHandler
     public void onEconomyCheck(AccountCheckEvent event) {
         //Try to see if we have an account for this UUID
-        dev.tycho.stonks.model.Account account = databaseManager.getAccountWithUUID(event.getAccount());
+        dev.tycho.stonks.model.core.Account account = databaseManager.getAccountWithUUID(event.getAccount());
         if (account != null) {
             event.hasAccount(true);
         }
@@ -157,7 +157,7 @@ public class ShopManager extends SpigotModule {
 
     @EventHandler
     public void onCurrencyAdd(PreCurrencyAddEvent event) {
-        dev.tycho.stonks.model.Account account = databaseManager.getAccountWithUUID(event.getTarget());
+        dev.tycho.stonks.model.core.Account account = databaseManager.getAccountWithUUID(event.getTarget());
         if (account != null) {
             event.setCancelled(true);
 
@@ -200,7 +200,7 @@ public class ShopManager extends SpigotModule {
 
     @EventHandler
     public void onCurrencySubtract(PreCurrencySubtractEvent event) {
-        dev.tycho.stonks.model.Account account = databaseManager.getAccountWithUUID(event.getSender());
+        dev.tycho.stonks.model.core.Account account = databaseManager.getAccountWithUUID(event.getSender());
         if (account != null) {
             event.setCancelled(true);
             Stonks.newChain()
@@ -242,7 +242,7 @@ public class ShopManager extends SpigotModule {
     @EventHandler
     public void onAmountCheck(PreAmountCheckEvent event) {
         //If they try to check an account balance for one of our accounts cancel it
-        dev.tycho.stonks.model.Account account = databaseManager.getAccountWithUUID(event.getAccount());
+        dev.tycho.stonks.model.core.Account account = databaseManager.getAccountWithUUID(event.getAccount());
         if (account != null) {
             event.setCancelled(true);
         }
@@ -250,7 +250,7 @@ public class ShopManager extends SpigotModule {
 
     @EventHandler
     public void onCurrencyCheck(PreCurrencyCheckEvent event) {
-        dev.tycho.stonks.model.Account account = databaseManager.getAccountWithUUID(event.getAccount());
+        dev.tycho.stonks.model.core.Account account = databaseManager.getAccountWithUUID(event.getAccount());
         if (account != null) {
             if (account.getTotalBalance() >= event.getAmountSent().doubleValue()) {
                 event.setHasEnough(true);
