+++ src/main/java/dev/tycho/stonks/model/core/AccountType.java
@@ -1,4 +1,4 @@
-package dev.tycho.stonks.model;
+package dev.tycho.stonks.model.core;
 
 public enum AccountType {
     CompanyAccount,
