+++ src/main/java/dev/tycho/stonks/model/core/Role.java
@@ -1,4 +1,4 @@
-package dev.tycho.stonks.model;
+package dev.tycho.stonks.model.core;
 
 public enum Role {
     CEO, Manager, Employee, Intern
