+++ src/main/java/dev/tycho/stonks/model/service/Service.java
@@ -1,0 +1,10 @@
+package dev.tycho.stonks.model.service;
+
+public class Service {
+  String name;
+  //days
+  long duration;
+
+
+
+}
