+++ src/main/java/dev/tycho/stonks/model/service/Subscriber.java
@@ -1,0 +1,5 @@
+package dev.tycho.stonks.model.service;
+
+public class Subscriber {
+
+}
